Move part of image_processing tests
